slight reorg for airflow
