--- docs/src/assets/temp.text
@@ -1,1 +1,0 @@
-text
